cleanup(02): physicsworld step 0.3 - replace wake island union-find
Replaces the local WakePointJointIsland find/union lambdas with the shared DisjointSet helper while preserving the existing parent/rank buffer setup order and equal-rank first-root tie-break.

Adjusts DisjointSet to borrow the existing uint8_t rank buffers used by PhysicsWorld instead of changing storage.

Validation: cmd.exe /c tools\validate_physics.bat passed; physics_regression_solver.csv matched baseline byte-exact (20001 lines), with 0 warnings and 0 errors. Log: Agentic\Logs\cleanup-02-step-0.3-validate-physics.log.

Comment audit: inspected touched source-bearing files SkullbonezSource/Physics/DisjointSet.h and SkullbonezSource/Physics/PhysicsWorld.cpp; 0 deferred.

+++ SkullbonezSource/Physics/DisjointSet.h
@@ -32,6 +32,7 @@ Related:
 
 #include <algorithm>
 #include <cstddef>
+#include <cstdint>
 #include <vector>
 
 namespace SkullbonezCore::Physics
@@ -40,7 +41,7 @@ namespace SkullbonezCore::Physics
 class DisjointSet
 {
 public:
-    DisjointSet( std::vector<int>& parent, std::vector<int>& rank, int count )
+    DisjointSet( std::vector<int>& parent, std::vector<uint8_t>& rank, int count )
         : m_parent( parent ), m_rank( rank ), m_count( count )
     {
     }
@@ -101,7 +102,7 @@ public:
 
 private:
     std::vector<int>& m_parent;
-    std::vector<int>& m_rank;
+    std::vector<uint8_t>& m_rank;
     int m_count;
 };
